Factoring out util functions
